VIRO-57 [skeletal]: properly concatenate bone transforms
This commit fixes three bugs that were breaking
skeletal animations, two simple, one big. The
simple ones are:

1. Duration was specified in seconds instead of
milliseconds.

2. We were interpolating matrices using the wrong
index (i instead of j).

The big one is we were concatenating bone
transform matrices incorrectly. For one, we were
using *local* transforms for bone matrices, when
FBX exports fully concatenated transform matrices
for bones. So we no longer have to do this ourselves
in the skinner.

Secondly, we had numerous coordinate space
issues. There are 3 coordinate spaces involved in
skinning: model space, bone-local space, and world
space. There are also 3 positions a matrix can be
in: original position, bind position, and animated
position. This is all fixed now, and heavily documented
in VROSkinner.h.

+++ ViroRenderer/VROSkeleton.h
@@ -30,7 +30,7 @@ public:
     virtual ~VROSkeleton() {}
     
     int getNumBones() const {
-        return _bones.size();
+        return (int)_bones.size();
     }
     const std::shared_ptr<VROBone> getBone(int index) const {
         return _bones[index];
